requirements.txt
aiohttp
requests
fastapi
redis
aioredis
fastapi-cache2[redis]
uvicorn[standart]
python-dotenv
